+++ CHANGELOG.md
@@ -11,5 +11,6 @@ CHANGELOG
 # Features
 - Redesign /match, /lastmatch and /recents output
 - Add more views to /match and buttons to switch between them
+- Use decorators to dry up logic with arguments for certain commands
 
 # Commands